+++ testing/src/main/java/io/github/kaktushose/jdac/testing/TestScenario.java
@@ -6,10 +6,11 @@ import com.github.kaktushose.jda.commands.definitions.description.ClassFinder;
 import com.github.kaktushose.jda.commands.definitions.interactions.InteractionDefinition;
 import com.github.kaktushose.jda.commands.definitions.interactions.command.CommandDefinition;
 import com.github.kaktushose.jda.commands.dispatching.instance.InteractionControllerInstantiator;
-import io.github.kaktushose.jdac.testing.invocation.ButtonInvocation;
-import io.github.kaktushose.jdac.testing.invocation.EntitySelectInvocation;
-import io.github.kaktushose.jdac.testing.invocation.SlashCommandInvocation;
-import io.github.kaktushose.jdac.testing.invocation.StringSelectInvocation;
+import io.github.kaktushose.jdac.testing.invocation.commands.SlashCommandInvocation;
+import io.github.kaktushose.jdac.testing.invocation.components.ButtonInvocation;
+import io.github.kaktushose.jdac.testing.invocation.components.EntitySelectInvocation;
+import io.github.kaktushose.jdac.testing.invocation.components.StringSelectInvocation;
+import io.github.kaktushose.jdac.testing.invocation.internal.Invocation;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.hooks.IEventManager;
