+++ testing/src/main/java/io/github/kaktushose/jdac/testing/invocation/ModalInvocation.java
@@ -1,6 +1,7 @@
 package io.github.kaktushose.jdac.testing.invocation;
 
 import io.github.kaktushose.jdac.testing.TestScenario.Context;
+import io.github.kaktushose.jdac.testing.invocation.internal.Invocation;
 import net.dv8tion.jda.api.events.interaction.ModalInteractionEvent;
 import net.dv8tion.jda.api.interactions.InteractionType;
 import net.dv8tion.jda.api.interactions.modals.ModalMapping;
