+++ testing/src/main/java/io/github/kaktushose/jdac/testing/reply/ModalEventReply.java
@@ -1,8 +1,9 @@
 package io.github.kaktushose.jdac.testing.reply;
 
 import io.github.kaktushose.jdac.testing.TestScenario.Context;
-import io.github.kaktushose.jdac.testing.invocation.Invocation;
+import io.github.kaktushose.jdac.testing.invocation.internal.Invocation;
 import io.github.kaktushose.jdac.testing.invocation.ModalInvocation;
+import io.github.kaktushose.jdac.testing.reply.internal.EventReply;
 import net.dv8tion.jda.api.interactions.modals.Modal;
 
 public final class ModalEventReply extends EventReply {
